Move closureConvert and validate compiler phases into separate files.
CHANGELOG_BEGIN
CHANGELOG_END
